backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pinecone-client==3.0.0
openai==1.58.1
python-multipart==0.0.6
